requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-bootstrap